Converter: 7fed83 - 128 new successes, 6 new failures
New successes:
@antv/g2
@asymmetrik/leaflet-d3
@atlaskit/feedback-collector
@babel/standalone
@cyberblast/config
@cyberblast/logger
@cyberblast/webserver
@edtr-io/mathquill
@firebase/auth-interop-types
@firebase/component
@jupyterlab/csvviewer-extension
@jupyterlab/documentsearch
@jupyterlab/documentsearch-extension
@jupyterlab/notebook
@material-ui/styles
@material-ui/system
@material-ui/types
@mdx-js/react
@newrelic/winston-enricher
@pager/jackrabbit
@pulumi/aws
@pusher/chatkit-client
@rdfjs/namespace
@rdfjs/parser-jsonld
@rdfjs/parser-n3
@rdfjs/serializer-jsonld
@rdfjs/serializer-jsonld-ext
@rebass/forms
@riotjs/dom-bindings
@theme-ui/components
@ungap/create-content
@ungap/global-this
@ungap/weakmap
akamai-edgeworkers
alpha-bravo
ant-design-pro
arg
aws-sdk
bcp-47
bcp-47-match
bizcharts
blake2
bmp-js
broccoli-output-wrapper
canvaskit-wasm
cassandra-store
change-case-object
clownface
cypress-image-snapshot
datadog-winston
datastore-core
datastore-level
diva.js
express-list-endpoints
fastbitset
fs-merger
function-bind
fundamental-react
googleapis
highlight-words-core
interface-datastore
it-all
json-form-data
jsreport-chrome-pdf
knuth-shuffle
kompression
lang.js
leaflet-editable
leaflet-label
level-ttl
lexicographic-integer
libphonenumber-js
mitsobox
native-url
node-loggly-bulk
node-object-hash
openssi-websdk
password
php-serialize
pica
port-numbers
poster-image
promise-abortable
promise-nodeify
purifycss-webpack
pvtsutils
ramda
range_check
rdf-dataset-ext
rdf-ext
react-cookies
react-countdown-circle-timer
react-fade-in
react-hook-mousetrap
react-imgix
react-mathquill
react-mic
react-native-share-extension
react-router-guard
react-stars
react-swipeable-views-utils
rfdc
rn-app-upgrade
scroll-into-view-if-needed
scss-parser
simple-statistics
std
sticky-session
strman
subleveldown
tape-catch
tape-promise
to-json-schema
trashable
tree-changes
truncate-middle
unstated-next
use-media-antd-query
use-merge-value
use-position
use-resize-observer
user-agents
vega-event-selector
vega-expression
vega-scale
vscode-windows-registry
windows-mutex
worker-plugin

New failures:
@material-ui/lab
p-some
stream-csv-as-json
stream-json
webdriverio
workbox-sw

Old failures
@aws-sdk/client-s3-browser
@aws-sdk/fetch-http-handler
@esfx/type-model
@feathersjs/authentication-oauth2
@ionic/react
@jimp/utils
@wordpress/components
activex-libreoffice
add-matchers
adone
asmcrypto.js
azure-mobile-apps
babylonjs
badgen
bunyan-format
c3
carbon-components-react
child-process-promise
combined-stream
csv-parse
d3-cloud
devextreme
fastify-plugin
forge-viewer
immutability-helper
jointjs
jsnlog
kafkajs
lunr
mangopay2-nodejs-sdk
mergerino
meteor-typings
minapp-env
monocle-ts
msportalfx-mock
multiplexjs
mz
node-notifier
node-vault
openfin
pixi.js
powerbi-visuals-tools
prismjs
rbx
react-md
redux-orm
redux-promise-middleware
rethinkdb
samchon-library
steamid
tabris
typescript-stl
ui-box
vuelidate
x-ray
xstream

+++ t/tensorflow__tfjs-core/src/main/scala/typingsSlinky/atTensorflowTfjsDashCore/distTypesMod/package.scala
@@ -6,7 +6,7 @@ import scala.scalajs.js.annotation._
 
 package object distTypesMod {
   type BackendValues = scala.scalajs.js.typedarray.Float32Array | scala.scalajs.js.typedarray.Int32Array | scala.scalajs.js.typedarray.Uint8Array | js.Array[scala.scalajs.js.typedarray.Uint8Array]
-  type DataValues = js.Array[String] | scala.scalajs.js.typedarray.Int32Array | scala.scalajs.js.typedarray.Float32Array | scala.scalajs.js.typedarray.Uint8Array
+  type DataValues = js.Array[String] | scala.scalajs.js.typedarray.Float32Array | scala.scalajs.js.typedarray.Int32Array | scala.scalajs.js.typedarray.Uint8Array
   type FlatVector = (js.Array[Boolean | Double]) | TypedArray
   type RegularArray[T] = js.Array[(js.Array[(js.Array[(js.Array[(js.Array[js.Array[T] | T]) | T]) | T]) | T]) | T]
   type ScalarLike = Double | Boolean | String | scala.scalajs.js.typedarray.Uint8Array
